pyproject.toml
[tool.poetry]
name = "gpto1"
version = "0.1.0"
description = ""
authors = ["Han Lee <lee.hanchung@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.10"
openai = "^1.47.0"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
tqdm = "^4.66.5"
matplotlib = "^3.9.2"
scipy = "^1.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
